pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "jinja2>=3.1.6",
    "markdown>=3.7",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.0",
]
